Initial work on map and associated files.
Several related files are found here.

Index.html is the page 'backbone', housing a div which will in turn
house the interactive map. The map itself is world.html, which uses
countryData.min.js to pull coordinate data to create an interactive
layer.

Miscellaneous files (formatting.css, layout.html) are also added for
completeness though are not required/linked to the primary files.